Rename and update the working code of R peaks Detection & Heart rate calculation
